quantum computing/backend/requirements.txt
fastapi
uvicorn[standard]
qiskit
requests
pydantic
python-dotenv
